requirements.txt
apache-airflow==2.7.1
sqlalchemy
fnmatch
python-dotenv